Day5_Prototype_to_Production/requirements.txt
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

